+++ spec/dummy/demo_app/app/controllers/lg_controller.rb
@@ -4,6 +4,8 @@ module DemoApp
   class LgController < ApplicationController
     layout false
 
+    focus_ring :status, :commits, :files, :diff
+
     def show
       render :show, palette: command_palette
     end